fix(gentoo): update package requirements for ohnyzsh setup
This commit updates the package requirements for setting up Oh My Zsh. Previously, the name parameter in the package task was using the variable package_requirements, but now it dynamically selects the appropriate package requirements based on the os_family Ansible fact. If the os_family is Gentoo, the package requirements are set to dev-vcs/git and app-shells/zsh. Otherwise, the default package requirements remain unchanged, consisting of git and zsh.

This change ensures that the correct packages are installed for setting up Oh My Zsh on different operating systems, improving compatibility and maintainability.